Register listeners only with both call and kick
The vhost-user protocol doesn't impose an order for the SET_VRING_KICK
and SET_VRING_CALL messages, which implies it's valid to emit the
first before the latter.

With the current code, this means that if the VMM sends SET_VRING_KICK
before SET_VRING_CALL, and the guest has already placed a request into
the vring, we may miss signaling the guest as that request may be
processed before we have an EventFd for "call".

To fix this, delay listener registration until we have an EventFd for
both call and kick, using "vring.queue.ready" as an indicator that the
vring has not been initialized yet.

Signed-off-by: Sergio Lopez <slp@redhat.com>